+++ tests/test_graph.py
@@ -93,12 +93,12 @@ class TestGraph(unittest.TestCase):
                                  original_node_1_link_weights[index])
             self.assertLessEqual(weight_difference, MAX_NOISE_WEIGHT)
 
-    def test_find_node_by_name(self):
-        found_node = self.test_graph.find_node_by_name('Node 2')
+    def test_find_node_by_value(self):
+        found_node = self.test_graph.find_node_by_value('Node 2')
         self.assertEqual(found_node, self.node_2)
 
-    def test_find_node_by_name_with_invalid_name(self):
-        self.assertIsNone(self.test_graph.find_node_by_name('bogus name'))
+    def test_find_node_by_value_with_invalid_value(self):
+        self.assertIsNone(self.test_graph.find_node_by_value('bogus value'))
 
     def test_remove_node(self):
         self.test_graph.remove_node(self.node_2)
@@ -113,22 +113,22 @@ class TestGraph(unittest.TestCase):
         self.assertEqual(self.test_graph.node_list,
                          [self.node_1, self.node_2, self.node_3])
 
-    def test_remove_node_by_name(self):
-        self.test_graph.remove_node_by_name('Node 1')
+    def test_remove_node_by_value(self):
+        self.test_graph.remove_node_by_value('Node 1')
         self.assertEqual(self.test_graph.node_list,
                          [self.node_2, self.node_3])
-        # Test that no links point to any node named 'Node 1'
+        # Test that no links point to any node valued 'Node 1'
         for n in self.test_graph.node_list:
-            self.assertFalse('Node 1' in [l.target.name for l in n.link_list])
+            self.assertFalse('Node 1' in [l.target.value for l in n.link_list])
 
-    def test_remove_node_by_name_with_invalid_name(self):
-        self.test_graph.remove_node_by_name('bogus name')
+    def test_remove_node_by_value_with_invalid_value(self):
+        self.test_graph.remove_node_by_value('bogus value')
         self.assertEqual(self.test_graph.node_list,
                          [self.node_1, self.node_2, self.node_3])
 
-    def test_has_node_with_name(self):
-        self.assertTrue(self.test_graph.has_node_with_name('Node 1'))
-        self.assertFalse(self.test_graph.has_node_with_name('bogus name'))
+    def test_has_node_with_value(self):
+        self.assertTrue(self.test_graph.has_node_with_value('Node 1'))
+        self.assertFalse(self.test_graph.has_node_with_value('bogus value'))
 
     def test_pick_with_no_starting_or_current_node(self):
         # Pick with starting_node=None and
@@ -169,49 +169,49 @@ class TestGraph(unittest.TestCase):
             self.assertEqual(len(n.link_list), 1)
             self.assertEqual(n.link_list[0].weight, 1)
         # Test the specifics of the graph
-        # Name of the node
-        # and name of the node it is linked to
-        self.assertEqual(built_graph.node_list[0].name, 'I')
-        self.assertEqual(built_graph.node_list[0].link_list[0].target.name,
+        # Value of the node
+        # and value of the node it is linked to
+        self.assertEqual(built_graph.node_list[0].value, 'I')
+        self.assertEqual(built_graph.node_list[0].link_list[0].target.value,
                          'have')
         # etc...
-        self.assertEqual(built_graph.node_list[1].name, 'have')
-        self.assertEqual(built_graph.node_list[1].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[1].value, 'have')
+        self.assertEqual(built_graph.node_list[1].link_list[0].target.value,
                          'nothing to say,.;!?:\\/\'"()[')
-        self.assertEqual(built_graph.node_list[2].name,
+        self.assertEqual(built_graph.node_list[2].value,
                          'nothing to say,.;!?:\\/\'"()[')
-        self.assertEqual(built_graph.node_list[2].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[2].link_list[0].target.value,
                          'and')
-        self.assertEqual(built_graph.node_list[3].name, 'and')
-        self.assertEqual(built_graph.node_list[3].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[3].value, 'and')
+        self.assertEqual(built_graph.node_list[3].link_list[0].target.value,
                          'I')
-        self.assertEqual(built_graph.node_list[4].name, 'I')
-        self.assertEqual(built_graph.node_list[4].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[4].value, 'I')
+        self.assertEqual(built_graph.node_list[4].link_list[0].target.value,
                          'am')
-        self.assertEqual(built_graph.node_list[5].name, 'am')
-        self.assertEqual(built_graph.node_list[5].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[5].value, 'am')
+        self.assertEqual(built_graph.node_list[5].link_list[0].target.value,
                          'saying')
-        self.assertEqual(built_graph.node_list[6].name, 'saying')
-        self.assertEqual(built_graph.node_list[6].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[6].value, 'saying')
+        self.assertEqual(built_graph.node_list[6].link_list[0].target.value,
                          'it')
-        self.assertEqual(built_graph.node_list[7].name, 'it')
-        self.assertEqual(built_graph.node_list[7].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[7].value, 'it')
+        self.assertEqual(built_graph.node_list[7].link_list[0].target.value,
                          'and')
-        self.assertEqual(built_graph.node_list[8].name, 'and')
-        self.assertEqual(built_graph.node_list[8].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[8].value, 'and')
+        self.assertEqual(built_graph.node_list[8].link_list[0].target.value,
                          'that')
-        self.assertEqual(built_graph.node_list[9].name, 'that')
-        self.assertEqual(built_graph.node_list[9].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[9].value, 'that')
+        self.assertEqual(built_graph.node_list[9].link_list[0].target.value,
                          'is')
-        self.assertEqual(built_graph.node_list[10].name, 'is')
-        self.assertEqual(built_graph.node_list[10].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[10].value, 'is')
+        self.assertEqual(built_graph.node_list[10].link_list[0].target.value,
                          'poetry')
-        self.assertEqual(built_graph.node_list[11].name, 'poetry')
-        self.assertEqual(built_graph.node_list[11].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[11].value, 'poetry')
+        self.assertEqual(built_graph.node_list[11].link_list[0].target.value,
                          '.')
-        self.assertEqual(built_graph.node_list[12].name, '.')
+        self.assertEqual(built_graph.node_list[12].value, '.')
         # Wraps around to first word
-        self.assertEqual(built_graph.node_list[12].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[12].link_list[0].target.value,
                          'I')
 
     def test_from_string_with_default_weights_and_merge_same_words(self):
@@ -223,18 +223,18 @@ class TestGraph(unittest.TestCase):
         # each with exactly one link pointing to the following word
         # (last word wrapping to the first)
         self.assertEqual(len(built_graph.node_list), 11)
-        # Test that no two nodes have the same name
+        # Test that no two nodes have the same value
         for index, node in enumerate(built_graph.node_list):
             for other_index, other_node in enumerate(built_graph.node_list):
                 if other_index == index:
                     continue
-                self.assertTrue(other_node.name != node.name)
+                self.assertTrue(other_node.value != node.value)
         # Node for 'I' should have two links, one pointing to 'have' and
         # one pointing to 'am'
         self.assertEqual(len(built_graph.node_list[0].link_list), 2)
-        self.assertEqual(built_graph.node_list[0].link_list[0].target.name,
+        self.assertEqual(built_graph.node_list[0].link_list[0].target.value,
                          'have')
-        self.assertEqual(built_graph.node_list[0].link_list[1].target.name,
+        self.assertEqual(built_graph.node_list[0].link_list[1].target.value,
                          'am')
 
     def test_from_string_with_custom_weights_and_merging_same_words(self):
@@ -248,51 +248,51 @@ class TestGraph(unittest.TestCase):
         # each with exactly one link pointing to the following word
         # (last word wrapping to the first)
         self.assertEqual(len(built_graph.node_list), 11)
-        # Test that no two nodes have the same name
+        # Test that no two nodes have the same value
         for index, node in enumerate(built_graph.node_list):
             for other_index, other_node in enumerate(built_graph.node_list):
                 if other_index == index:
                     continue
-                self.assertTrue(other_node.name != node.name)
+                self.assertTrue(other_node.value != node.value)
 
         # Test the order in which nodes were added is as expected
-        self.assertEqual(built_graph.node_list[0].name,  'I')
-        self.assertEqual(built_graph.node_list[1].name,  'have')
-        self.assertEqual(built_graph.node_list[2].name,
+        self.assertEqual(built_graph.node_list[0].value,  'I')
+        self.assertEqual(built_graph.node_list[1].value,  'have')
+        self.assertEqual(built_graph.node_list[2].value,
                          'nothing to say,.;!?:\\/\'"()[')
-        self.assertEqual(built_graph.node_list[3].name,  'and')
-        self.assertEqual(built_graph.node_list[4].name,  'am')
-        self.assertEqual(built_graph.node_list[5].name,  'saying')
-        self.assertEqual(built_graph.node_list[6].name,  'it')
-        self.assertEqual(built_graph.node_list[7].name,  'that')
-        self.assertEqual(built_graph.node_list[8].name,  'is')
-        self.assertEqual(built_graph.node_list[9].name,  'poetry')
-        self.assertEqual(built_graph.node_list[10].name, '.')
+        self.assertEqual(built_graph.node_list[3].value,  'and')
+        self.assertEqual(built_graph.node_list[4].value,  'am')
+        self.assertEqual(built_graph.node_list[5].value,  'saying')
+        self.assertEqual(built_graph.node_list[6].value,  'it')
+        self.assertEqual(built_graph.node_list[7].value,  'that')
+        self.assertEqual(built_graph.node_list[8].value,  'is')
+        self.assertEqual(built_graph.node_list[9].value,  'poetry')
+        self.assertEqual(built_graph.node_list[10].value, '.')
 
-        # Exhaustively test the node with the name 'I'
-        # Find the node with the name 'I' (do this manually)
-        i_node = next(n for n in built_graph.node_list if n.name == 'I')
+        # Exhaustively test the node with the value 'I'
+        # Find the node with the value 'I' (do this manually)
+        i_node = next(n for n in built_graph.node_list if n.value == 'I')
         # Compare link contents of 'I' against expected values
         self.assertEqual(len(i_node.link_list), 5)
-        self.assertEqual(i_node.link_list[0].target.name, 'and')
+        self.assertEqual(i_node.link_list[0].target.value, 'and')
         self.assertEqual(i_node.link_list[0].weight, 1 + 5)
-        self.assertEqual(i_node.link_list[1].target.name, 'I')
+        self.assertEqual(i_node.link_list[1].target.value, 'I')
         self.assertEqual(i_node.link_list[1].weight, 2 + 5 + 2)
-        self.assertEqual(i_node.link_list[2].target.name, 'have')
+        self.assertEqual(i_node.link_list[2].target.value, 'have')
         self.assertEqual(i_node.link_list[2].weight, 3)
-        self.assertEqual(i_node.link_list[3].target.name, '.')
+        self.assertEqual(i_node.link_list[3].target.value, '.')
         self.assertEqual(i_node.link_list[3].weight, 1)
-        self.assertEqual(i_node.link_list[4].target.name, 'am')
+        self.assertEqual(i_node.link_list[4].target.value, 'am')
         self.assertEqual(i_node.link_list[4].weight, 3)
 
     def test_from_string_with_punc_inside_word(self):
         source = ("I'm a human!")
         built_graph = Graph.from_string(source)
         self.assertEqual(len(built_graph.node_list), 4)
-        self.assertEqual(built_graph.node_list[0].name,  "I'm")
-        self.assertEqual(built_graph.node_list[1].name,  'a')
-        self.assertEqual(built_graph.node_list[2].name,  'human')
-        self.assertEqual(built_graph.node_list[3].name,  '!')
+        self.assertEqual(built_graph.node_list[0].value,  "I'm")
+        self.assertEqual(built_graph.node_list[1].value,  'a')
+        self.assertEqual(built_graph.node_list[2].value,  'human')
+        self.assertEqual(built_graph.node_list[3].value,  '!')
 
     def test_from_file_is_same_as_from_string_of_file_contents(self):
         # Defaults:
@@ -312,18 +312,18 @@ class TestGraph(unittest.TestCase):
         graph_from_string = Graph.from_string(source_as_string)
         for file_node, string_node in zip(graph_from_file.node_list,
                                           graph_from_string.node_list):
-            self.assertEqual(file_node.name, string_node.name)
+            self.assertEqual(file_node.value, string_node.value)
             for file_link, string_link in zip(file_node.link_list,
                                               string_node.link_list):
-                self.assertEqual(file_link.target.name,
-                                 string_link.target.name)
+                self.assertEqual(file_link.target.value,
+                                 string_link.target.value)
                 self.assertEqual(file_link.weight, string_link.weight)
 
     def test_from_string_with_default_group_marker(self):
         source = ('I have <<nothing to say,.;!?:\\/\'"()[>>'
                   'and I am saying it and that is poetry.')
         built_graph = Graph.from_string(source)
-        self.assertEqual(built_graph.node_list[2].name,
+        self.assertEqual(built_graph.node_list[2].value,
                          'nothing to say,.;!?:\\/\'"()[')
 
     def test_from_string_with_custom_group_marker(self):
@@ -332,7 +332,7 @@ class TestGraph(unittest.TestCase):
         built_graph = Graph.from_string(source,
                                         group_marker_opening='{',
                                         group_marker_closing='}')
-        self.assertEqual(built_graph.node_list[2].name,
+        self.assertEqual(built_graph.node_list[2].value,
                          'nothing to say,.;!?:\\/\'"()[')
 
     def test_from_file_with_custom_group_marker(self):
@@ -344,5 +344,5 @@ class TestGraph(unittest.TestCase):
         graph_from_file = Graph.from_file(file_name,
                                           group_marker_opening='{',
                                           group_marker_closing='}')
-        self.assertEqual(graph_from_file.node_list[2].name,
+        self.assertEqual(graph_from_file.node_list[2].value,
                          'nothing to say,.;!?:\\/\'"()[')